requirements.txt
amqp==2.1.4
appnope==0.1.0
billiard==3.5.0.2
celery==4.0.2
chainer==1.20.0
cycler==0.10.0
decorator==4.0.10
Django==1.10.5
djangorestframework==3.5.3
filelock==2.0.7
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
kombu==4.0.2
matplotlib==2.0.0
nose==1.3.7
numpy==1.12.0
olefile==0.44
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
protobuf==3.1.0.post1
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
redis==2.10.5
requests==2.12.4
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.4
traitlets==4.3.1
vine==1.1.3
wcwidth==0.1.7
